+++ R/helpers.R
@@ -1,7 +1,0 @@
-#' myhelpers.
-#'
-#' @name myhelpers
-#' @docType package
-#' @import colorspace maps geocors ncdf4 reshape2
-#'
-NULL