+++ R/myhelpers.R
@@ -2,5 +2,5 @@
 #'
 #' @name myhelpers
 #' @docType package
-#' @import maps ncdf4 reshape2 colorspace geocors
+#' @import maps ncdf4 reshape2 colorspace geocors lubridate
 NULL